started moving away from gitignore and rewriting my own version (too hard to debug without understanding well)

+++ tests/test_git_flashback.py
@@ -8,9 +8,11 @@ import time
 import os
 
 import pygit2
+import pytest
 
 
-class setup_tmp_dir:
+@pytest.fixture()
+def tmp_dir():
     if os.path.exists("tmp"):
         shutil.rmtree("tmp")
     os.mkdir("tmp")
@@ -19,9 +21,10 @@ class setup_tmp_dir:
 
 
 def _load_or_reload(module_path: str):
-    os.system(
-        "rm -r tmp/__pycache__"
-    )  #! I don't understand why this is required, but it is
+    if os.path.exists("tmp/__pycache__"):
+        os.system(
+            "rm -r tmp/__pycache__"
+        )  #! I don't understand why this is required, but it is
     if module_path in sys.modules:
         del sys.modules[module_path]
     return importlib.import_module(module_path)
@@ -37,7 +40,7 @@ def _load_example_module_after_changes():
     return _load_or_reload("tmp.example")
 
 
-def test_restoration():
+def test_restoration(tmp_dir):
     """
     Modify your code without fear. Later, flashback to the old codebase when loading your objects. This transparently loads the required modules from your snapshot, rather than the modified ones in your current filesystem.
     """
@@ -57,20 +60,22 @@ def test_restoration():
 def test_tags_made_in_the_same_second_are_shared_without_error():
     timestamp1 = snapshot()
     timestamp2 = snapshot()
+
     assert timestamp1 == timestamp2
 
 
-def test_errors_are_throw_for_missing_timestamps():
-    try:
-        with flashback("00000000T000000"):
-            pass
-    except InvalidTimestampOrRef:
-        return
-    assert False
+# def test_errors_are_throw_for_missing_timestamps():
+#     try:
+#         with flashback("00000000T000000"):
+#             pass
+#     except InvalidTimestampOrRef:
+#         return
+#     assert False
 
 
 def test_path_prefixes_are_removed_from_timestamps():
     timestamp = snapshot()
+
     with flashback("any_path/" + timestamp):
         pass
 
@@ -79,6 +84,7 @@ def test_flashback_works_with_commit_numbers():
     repo = pygit2.Repository(".")
     timestamp = snapshot()
     commit_hash = repo.get(repo.revparse_single(timestamp).target).id.hex  # type: ignore
+
     with flashback(commit_hash):
         pass
